+++ lib/cuda/cudaUpchannelizer_chord_U128.cpp
@@ -689,7 +689,7 @@ cudaEvent_t cudaUpchannelizer_chord_U128::execute(cudaPipelineState& /*pipestate
     if (err != CUDA_SUCCESS) {
         const char* errStr;
         cuGetErrorString(err, &errStr);
-        ERROR("cuLaunchKernel: Error number: {}: {}", err, errStr);
+        ERROR("cuLaunchKernel: Error number: {}: {}", (int)err, errStr);
     }
 
     // Copy results back to host memory
